refactor : new components ProductListHeader, FilterModal
