+++ src/pages/ProductList/ProductList.scss
@@ -13,106 +13,6 @@
     cursor: pointer;
   }
 
-  .header {
-    position: sticky;
-    top: 0;
-    background-color: $bg-light;
-    border-top: 1px $border solid;
-    border-bottom: 1px $border solid;
-    z-index: 1000;
-
-    .headerTitle {
-      max-width: 1170px;
-      margin: 0 auto;
-      padding: 15px 15px 16px;
-      font-family: 'Source Sans Pro', sans-serif;
-
-      .breadCrumbs {
-        font-size: 14px;
-        font-weight: 300;
-
-        .breadCrumbsItem::after {
-          content: ' / ';
-        }
-
-        .breadCrumbsItem:last-child::after {
-          content: '';
-        }
-      }
-
-      h1 {
-        font-size: 19px;
-        font-weight: 700;
-      }
-    }
-
-    .filterBtn {
-      box-sizing: content-box;
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      width: 150px;
-      height: 100%;
-      padding: 0 20px;
-      position: absolute;
-      top: 50%;
-      right: 0;
-      border-top: 1px $border solid;
-      border-bottom: 1px $border solid;
-      border-left: 1px $border solid;
-      background-color: #fff;
-      transform: translateY(-50%);
-      z-index: 100;
-
-      em {
-        pointer-events: none;
-      }
-
-      &.active {
-        .fa-plus {
-          transform: rotate(-45deg);
-          transition: 200ms ease;
-        }
-      }
-
-      .fa-plus {
-        font-size: 16px;
-        transition: 200ms ease;
-      }
-
-      .filterModal {
-        visibility: hidden;
-        opacity: 0;
-        position: absolute;
-        top: 64px;
-        right: 0;
-        width: calc(100% + 1px);
-        margin-top: 6px;
-        padding: 5px 0 6px;
-        background-color: #fff;
-        border: 1px $border solid;
-        border-top: none;
-        border-right: none;
-        text-align: start;
-        font-family: 'Source Sans Pro', sans-serif;
-        font-weight: 300;
-        transition: 200ms ease;
-
-        &.active {
-          visibility: visible;
-          opacity: 1;
-          margin-top: 0;
-          transition: 200ms ease;
-        }
-
-        li {
-          padding-bottom: 7px;
-          padding-left: 20px;
-        }
-      }
-    }
-  }
-
   .list {
     max-width: 1170px;
     margin: 0 auto;
