+++ src/pages/ProductList/ProductListHeader/ProductListHeader.scss
@@ -1,0 +1,73 @@
+@import '../../../styles/constant.scss';
+@import url('https://fonts.googleapis.com/css2?family=Cutive&family=Source+Sans+Pro:wght@200;400;700&display=swap');
+
+.ProductList {
+  .header {
+    position: sticky;
+    top: 0;
+    background-color: $bg-light;
+    border-top: 1px $border solid;
+    border-bottom: 1px $border solid;
+    z-index: 1000;
+
+    .headerTitle {
+      max-width: 1170px;
+      margin: 0 auto;
+      padding: 15px 15px 16px;
+      font-family: 'Source Sans Pro', sans-serif;
+
+      .breadCrumbs {
+        font-size: 14px;
+        font-weight: 300;
+
+        .breadCrumbsItem::after {
+          content: ' / ';
+        }
+
+        .breadCrumbsItem:last-child::after {
+          content: '';
+        }
+      }
+
+      h1 {
+        font-size: 19px;
+        font-weight: 700;
+      }
+    }
+
+    .filterBtn {
+      box-sizing: content-box;
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      width: 150px;
+      height: 100%;
+      padding: 0 20px;
+      position: absolute;
+      top: 50%;
+      right: 0;
+      border-top: 1px $border solid;
+      border-bottom: 1px $border solid;
+      border-left: 1px $border solid;
+      background-color: #fff;
+      transform: translateY(-50%);
+      z-index: 100;
+
+      em {
+        pointer-events: none;
+      }
+
+      &.active {
+        .fa-plus {
+          transform: rotate(-45deg);
+          transition: 200ms ease;
+        }
+      }
+
+      .fa-plus {
+        font-size: 16px;
+        transition: 200ms ease;
+      }
+    }
+  }
+}
